requirements.txt
joblib
mljar-supervised
aiogram==2.14.3
asyncio
cloudpickle
numpy
scikit-learn
pandas
python-dotenv


